+++ include/H2DE/H2DE.h
@@ -1,4 +1,4 @@
 /**
- * Version: 2.0.21 */
+ * Version: 2.0.22 */
 
 #include <H2DE_engine.h>